+++ src/PriceObserver.Web.Api/Extensions/ServerCallContextExtensions.cs
@@ -5,11 +5,11 @@ namespace PriceObserver.Web.Api.Extensions;
 
 public static class ServerCallContextExtensions
 {
-    public static long GetUserId(this ServerCallContext context)
+    public static int GetUserId(this ServerCallContext context)
     {
         var user = context.GetHttpContext().User;
         var userId = user.Claims.Single(x => x.Type == ClaimTypes.NameIdentifier).Value;
 
-        return long.Parse(userId);
+        return int.Parse(userId);
     }
 }
